Modul for the Background Colour Changer menu
It is a .js,.php and .HTML file in this folder. Aswell as a description of the code in English and Ukrainian. If you have any problems with the Code ask ChatGOT to explain/integrate it as it can do a far better job at doing that than me (lol). The code is mostly ready to use, just adjust the global variables to fit with the other js and HTML scripts (if there really are any, idk). The PHP needs adjustment. Also it sets a background colour not and image but that's an easy fix in the (not included) css file by setting the e.g "background-colour" to "background-image". As I said just ask ChatGPT, he smarter than all of us together. For actually using the code, just put it out of the folder and put it in the current sctrcture (I can't do that because I don't know it). If you have problems understanding the English use DeepL (https://www.deepl.com/translator). Have a nice day :)